+++ mcp/packages/server/src/PenpotMcpServer.ts
@@ -23,10 +23,12 @@ import { FileCreateTool } from "./tools/FileCreateTool";
 import { FileBindContextTool, FileGetContextTool, FileReleaseContextTool } from "./tools/FileContextTools";
 import { PageCreateTool, PageListTool, PageRenameTool, PageSetCurrentTool } from "./tools/PageTools";
 import {
+    ShapeDeleteTool,
     ShapeCreateFrameTool,
     ShapeCreateImageTool,
     ShapeCreateRectTool,
     ShapeCreateTextTool,
+    ShapeUpdateTool,
 } from "./tools/ShapeCreateTools";
 import { PenpotRpcClient } from "./PenpotRpcClient";
 import { ReplServer } from "./ReplServer";
@@ -230,6 +232,8 @@ export class PenpotMcpServer {
             new ShapeCreateRectTool(this),
             new ShapeCreateTextTool(this),
             new ShapeCreateImageTool(this),
+            new ShapeUpdateTool(this),
+            new ShapeDeleteTool(this),
             new ExecuteCodeTool(this),
             new HighLevelOverviewTool(this),
             new PenpotApiInfoTool(this, this.apiDocs),